requirements.txt
streamlit
langchain
langchain_cohere
python-dotenv
pandas
langchain_community
requests